requirements.txt
APScheduler
boto3
click
Flask
Flask-Cors
GitPython
gunicorn
io
itsdangerous
Jinja2
MarkupSafe
numpy
pandas
python-dateutil
pytz
six
Werkzeug
